add error handling in deleteID&updatePrice

+++ src/main/java/org/kieszek/rentalcars/RentalCarsService.java
@@ -33,11 +33,15 @@ public class RentalCarsService {
     }
 
     public void deleteId(Long id) {
+        boolean exist =rentalCarsRepository.existsById(id);
+        if (!exist){
+            throw new RuntimeException("This car not exist - "+id);
+        }
         rentalCarsRepository.deleteById(id);
     }
 
     public RentalCarsDto updatePrice(Long id, Integer price) {
-        RentalCars rentalCars = rentalCarsRepository.findById(id).orElseThrow();
+        RentalCars rentalCars = rentalCarsRepository.findById(id).orElseThrow(() -> new RuntimeException("Car not exist"));
         rentalCars.setPrice(price);
         RentalCars saved = rentalCarsRepository.save(rentalCars);
         return saved.toDto();
